+++ modules/control/run.py
@@ -213,7 +213,14 @@ def control_run(units: List[unit.Unit] = [], inputs: List[Image.Image] = [], ini
         num_units += 1
         debug(f'Control unit: i={num_units} type={u.type} enabled={u.enabled}')
         if not u.enabled:
+            shared.log.debug(f'Control unit disabled: i={num_units}')
+            if u.controlnet is not None and u.controlnet.model is not None:
+                shared.log.debug(f'Control unit offload: model="{u.controlnet.model_id}"')
+                sd_models.move_model(u.controlnet.model, devices.cpu)
             continue
+        else:
+            if u.controlnet is not None and u.controlnet.model is not None:
+                sd_models.move_model(u.controlnet.model, devices.device)
         if unit_type == 't2i adapter' and u.adapter.model is not None:
             active_process.append(u.process)
             active_model.append(u.adapter)