+++ CHANGELOG.md
@@ -75,6 +75,7 @@ To use and of the new models, simply select model from *Networks -> Reference* a
   *note*: requires reset of selected offload option  
 - control module allows reszing of indivudual process override images to match input image  
   for example: set size->before->method:nearest, mode:fixed or mode:fill  
+- automatically offload disabled controlnet units  
 - update **IPEX** to 2.1.40+xpu on Linux, thanks @Disty0!  
 - general **ROCm** fixes, thanks @lshqqytiger!  
 - fix full vae previews, thanks @Disty0!